fix: todos 상태 위치 이동 및 추가후 할일 목록 바로 미반영되는  이슈 처리
