+++ src/components/TodoList.tsx
@@ -3,41 +3,15 @@ import { BASE_URL } from '../api/url';
 import TodoItem from './TodoItem';
 import { TodoItemType } from '../types/index';
 
-const TodoList: React.FC = () => {
-  const [todos, setTodos] = useState<TodoItemType[]>([]);
+interface TodosProps {
+  todos: TodoItemType[];
+}
 
-  useEffect(() => {
-    const access_token = localStorage.getItem('accessToken');
-
-    const getTodos = async () => {
-      const response = await fetch(`${BASE_URL}/todos`, {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${access_token}`,
-        },
-      });
-
-      const todosData = await response.json();
-
-      if (response.ok) {
-        setTodos(todosData);
-      }
-
-      if (!response.ok) {
-        throw new Error(
-          todosData.message || '할일 목록을 정상적으로 가져오지 못했습니다.'
-        );
-      }
-    };
-
-    getTodos();
-  }, []);
-  // console.log('todos: ', todos);
-
+const TodoList: React.FC<TodosProps> = ({ todos }) => {
   return (
-    <>
+    <ul>
       <TodoItem todos={todos} />
-    </>
+    </ul>
   );
 };
