+++ src/pages/Todo.tsx
@@ -1,9 +1,46 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import TodoList from '../components/TodoList';
 import { BASE_URL } from '../api/url';
+import { TodoItemType } from '../types/index';
+import { useNavigate } from 'react-router-dom';
 
 const Todo = () => {
   const [todoText, setTodoText] = useState<string>('');
+  const [todos, setTodos] = useState<TodoItemType[]>([]);
+
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const access_token = localStorage.getItem('accessToken');
+
+    if (!access_token) {
+      navigate('/signin');
+      return;
+    }
+
+    const getTodos = async () => {
+      const response = await fetch(`${BASE_URL}/todos`, {
+        method: 'GET',
+        headers: {
+          Authorization: `Bearer ${access_token}`,
+        },
+      });
+
+      const todosData = await response.json();
+
+      if (response.ok) {
+        setTodos(todosData);
+      }
+
+      if (!response.ok) {
+        throw new Error(
+          todosData.message || '할일 목록을 정상적으로 가져오지 못했습니다.'
+        );
+      }
+    };
+
+    getTodos();
+  }, [todoText, navigate]);
 
   const submitHandler = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
@@ -30,9 +67,11 @@ const Todo = () => {
       setTodoText('');
     }
 
-    throw new Error(
-      resData.message || '할일이 정상적으로 추가되지 않았습니다.'
-    );
+    if (!response.ok) {
+      throw new Error(
+        resData.message || '할일이 정상적으로 추가되지 않았습니다.'
+      );
+    }
   };
 
   return (
@@ -49,7 +88,7 @@ const Todo = () => {
         </button>
       </form>
       <hr />
-      <TodoList />
+      <TodoList todos={todos} />
     </>
   );
 };
